service/Services/entities/src/main/java/org/infy/idp/entities/ge/jenkins/TestAnalysis.java
/***********************************************************************************************
*
* Copyright 2018 Infosys Ltd.
* Use of this source code is governed by MIT license that can be found in the LICENSE file or at
* https://opensource.org/licenses/MIT.
*
***********************************************************************************************/

package org.infy.idp.entities.ge.jenkins;

import com.google.gson.annotations.Expose;
import com.google.gson.annotations.SerializedName;

/**
 * Entity to store Test analysis details
 * 
 * @author Infosys
 *
 */
public class TestAnalysis {
	@SerializedName("Tool")
	@Expose
	private String tool;
	@SerializedName("TotalTests")
	@Expose
	private int totalTests;
	@SerializedName("Success")
	@Expose
	private int success;
	@SerializedName("Failure")
	@Expose
	private int failure;
	@SerializedName("status")
	@Expose
	private String status;

	public String getTool() {
		return tool;
	}

	public void setTool(String tool) {
		this.tool = tool;
	}

	public int getTotalTests() {
		return totalTests;
	}

	public void setTotalTests(int totalTests) {
		this.totalTests = totalTests;
	}

	public int getSuccess() {
		return success;
	}

	public void setSuccess(int success) {
		this.success = success;
	}

	public int getFailure() {
		return failure;
	}

	public void setFailure(int failure) {
		this.failure = failure;
	}

	public String getStatus() {
		return status;
	}

	public void setStatus(String status) {
		this.status = status;
	}

}
